CRUD groups and finishing app

+++ src/storage/group/createGroups.ts
@@ -1,15 +1,21 @@
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { GROUP_COLLECTION } from '@storage/storageConfig';
 import { getAllGroups } from '@storage/group/getAllGroups';
+import { AppError } from '@utils/AppError';
 
 export const createGroups = async (newGroup: string) => {
   try {
     const storedGroups = await getAllGroups();
 
+    const groupAlreadyExists = storedGroups.includes(newGroup);
+
+    if (groupAlreadyExists) {
+      throw new AppError('Grupo já existe');
+    }
+
     const storage = JSON.stringify([...storedGroups, newGroup]);
     await AsyncStorage.setItem(GROUP_COLLECTION, storage);
   } catch (error) {
-    console.error(error);
     throw error;
   }
 };
